prometheus_client/requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers
prometheus-client
prometheus-api-client==0.5.3
python-dotenv
datetime
